# shared/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Technomart API (auto-generated)",
    "version": "1.0.0",
    "description": "This OpenAPI document is generated from the existing JavaScript service layer. Update via `npm run generate:openapi` after modifying services or schemas."
  },
  "paths": {
...
    "/employees/{id}": {
      "delete": {
        "operationId": "employee_deleteEmployee",
        "summary": "Generated from src/api/services/employeeService.js#deleteEmployee",
        "tags": ["employeeService"],
        "responses": {
          "default": {
            "description": "Auto-generated response placeholder"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "operationId": "employee_updateEmployee",
        "summary": "Generated from src/api/services/employeeService.js#updateEmployee",
        "tags": ["employeeService"],
        "responses": {
          "default": {
            "description": "Auto-generated response placeholder"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Auto-generated placeholder body",
          "required": false
        }
      }
    },
...
  },
...
}